moving configurator into distro
