+++ opentelemetry-distro/setup.cfg
@@ -51,6 +51,8 @@ where = src
 [options.entry_points]
 opentelemetry_distro =
     otel_distro = opentelemetry.distro:OpenTelemetryDistro
+opentelemetry_configurator =
+    otel_configurator = opentelemetry.distro:Configurator
 
 [options.extras_require]
 test =
