+++ opentelemetry-sdk/setup.cfg
@@ -43,7 +43,6 @@ zip_safe = False
 include_package_data = True
 install_requires =
     opentelemetry-api == 0.17.dev0
-    opentelemetry-instrumentation == 0.17.dev0
 
 [options.packages.find]
 where = src
@@ -56,8 +55,6 @@ opentelemetry_tracer_provider =
 opentelemetry_exporter =
     console_span = opentelemetry.sdk.trace.export:ConsoleSpanExporter
     console_metrics = opentelemetry.sdk.metrics.export:ConsoleMetricsExporter
-opentelemetry_configurator =
-    sdk_configurator = opentelemetry.sdk.configuration:Configurator
 opentelemetry_ids_generator =
     random = opentelemetry.sdk.trace.ids_generator:RandomIdsGenerator
